Refactor out generateAndRunTests
