+++ lib/Generate.js
@@ -3,7 +3,10 @@
 const path = require("path"),
   spawn = require("cross-spawn"),
   fs = require("fs-extra"),
+  _ = require("lodash"),
+  Murmur = require("murmur-hash-js"),
   Version = require("./version.js"),
+  supportsColor = require("supports-color"),
   Runner = require("./runner.js"),
   pipeFilename = require("./pipe-filename.js");
 
@@ -41,9 +44,7 @@ function readUtf8(filepath) {
   });
 }
 
-function generateElmJson(pathToElmBinary /*:string*/, filePaths /*:Array<string>*/) {
-  const projectRootDir = path.resolve(Runner.findNearestElmPackageDir(filePaths));
-  const testRootDir = path.resolve(path.join(projectRootDir, "tests"));
+function generateElmJson(projectRootDir/*:string*/, testRootDir/*:string*/, pathToElmBinary /*:string*/, filePaths /*:Array<string>*/) {
   const generatedCodeDir = path.resolve(
     path.join(projectRootDir, "elm-stuff", "generated-code", "elm-explorations", "test")
   );
@@ -230,6 +231,125 @@ function askElmForMissingTransitiveDependencies(pathToElmBinary, pathtoElmProjec
   }
 }
 
+function generateMainModule(fuzz/*:number*/, seed/*:number*/, report/*:string*/, testFilePaths/*:Array<string>*/, tests/*:Array<Object>*/, generatedSrc/*:string*/, processes/*:number*/) {
+  // Building things like:
+  //
+  // import MyTests
+  //
+  // MyTests.suite
+  const imports = _.map(tests, function(test) {
+    return "import " + test.name;
+  });
+  const testList = _.map(tests, function(mod) {
+    return (
+      '    Test.describe "' +
+      mod.name +
+      '" [' +
+      _.map(mod.tests, function(test) {
+        return mod.name + "." + test;
+      }).join(",\n    ") +
+      "]"
+    );
+  });
 
-module.exports = {prepareCompiledJsFile, generateElmJson};
+  if (testList.length === 0) {
+    const errorMessage =
+      testFilePaths.length > 0
+        ? "I couldn't find any exposed values of type Test in files matching \"" +
+          testFilePaths.toString() +
+          '"\n\nMaybe try running elm-test with no arguments?'
+        : "I couldn't find any exposed values of type Test in any *.elm files in the tests/ directory of your project's root directory.\n\nTo generate some initial tests to get things going, run elm-test init";
+
+    console.error(errorMessage);
+    process.exit(1);
+  }
+
+  function sanitizedToString(str) {
+    return '"' + str.replace(/\\/g, "\\\\").replace(/"/g, '\\"') + '"';
+  }
+
+  function getReportCode() {
+    switch (report) {
+      case "json":
+        return "JsonReport";
+      case "junit":
+        return "JUnitReport";
+      default:
+        if (supportsColor) {
+          return "(ConsoleReport UseColor)";
+        } else {
+          return "(ConsoleReport Monochrome)";
+        }
+    }
+  }
+
+  if (isNaN(seed)) {
+    seed = Math.floor(Math.random() * (407199254740991)) + 1000;
+  }
+
+  var opts = {
+    fuzz: isNaN(fuzz) ? "Nothing" : "Just " + fuzz,
+    seed: seed,
+    report: getReportCode(),
+    paths: testFilePaths.map(sanitizedToString).join(",")
+  };
+
+  var optsCode =
+    "{ runs = " +
+    opts.fuzz +
+    ", report = " +
+    opts.report +
+    ", seed = " +
+    opts.seed +
+    ", processes = " +
+    processes +
+    ", paths = [" +
+    opts.paths +
+    "]}";
+
+  const testFileBody = [
+    imports.join("\n"),
+    "",
+    "import Test.Reporter.Reporter exposing (Report(..))",
+    "import Console.Text exposing (UseColor(..))",
+    "import Test.Runner.Node",
+    "import Test",
+    "",
+    "main : Test.Runner.Node.TestProgram",
+    "main =",
+    "    [ " + testList.join(", ") + " ]",
+    "        |> Test.concat",
+    "        |> Test.Runner.Node.run " + optsCode
+  ].join("\n");
+
+  // Generate a filename that incorporates the hash of file contents.
+  // This way, if you run e.g. `elm-test Foo.elm` and then `elm-test Bar.elm`
+  // and then re-run `elm-test Foo.elm` we still have a cached `Main` for
+  // `Foo.elm` (assuming none of its necessary imports have changed - and
+  // why would they?) so we don't have to recompile it.
+  const salt = Murmur.murmur3(testFileBody);
+  const moduleName = "Main" + salt;
+  const mainPath = path.join(generatedSrc, "Test", "Generated");
+  const mainFile = path.join(mainPath, moduleName + ".elm");
+  // We'll be putting the generated Main in something like this:
+  //
+  // my-project-name/elm-stuff/generated-code/elm-community/elm-test/src/Test/Generated/Main123456.elm
+  const testFileContents = [
+    "module Test.Generated." + moduleName + " exposing (main)",
+    testFileBody
+  ].join("\n\n");
+
+  // Make sure src/Test/Generated/ exists so we can write the file there.
+  fs.mkdirpSync(mainPath);
+
+  // Always write the file, in order to update its timestamp. This is important,
+  // because if we run `elm-make Main123456.elm` and that file's timestamp did
+  // not change, elm-make will short-circuit and not recompile *anything* - even
+  // if some of Main's dependencies (such as an individual test file) changed.
+  fs.writeFileSync(mainFile, testFileContents);
+
+  return mainFile;
+}
+
+module.exports = {prepareCompiledJsFile, generateElmJson, generateMainModule};
