+++ lib/elm-test.js
@@ -31,8 +31,6 @@ process.on("uncaughtException", function(error) {
 var compile = require("node-elm-compiler").compile,
   fs = require("fs-extra"),
   os = require("os"),
-  supportsColor = require("supports-color"),
-  murmur = require("murmur-hash-js"),
   glob = require("glob"),
   path = require("path"),
   util = require("util"),
@@ -168,9 +166,6 @@ else if (args._[0] == "init") {
   process.exit(0);
 }
 
-var projectRootDir;
-var testRootDir;
-
 function runTests(generatedCodeDir/*:string*/, testFile/*:string*/) {
   const dest = path.resolve(path.join(generatedCodeDir, "elmTestOutput.js"));
 
@@ -229,130 +224,7 @@ function resolveGlobs(fileGlobs) {
   return _.flatMap(globs, resolveFilePath);
 }
 
-function generateAndRunTests(tests, testFilePaths, generatedCodeDir, generatedSrc) {
-  // Building things like:
-  //
-  // import MyTests
-  //
-  // MyTests.suite
-  var imports = _.map(tests, function(test) {
-    return "import " + test.name;
-  });
-  var testList = _.map(tests, function(mod) {
-    return (
-      '    Test.describe "' +
-      mod.name +
-      '" [' +
-      _.map(mod.tests, function(test) {
-        return mod.name + "." + test;
-      }).join(",\n    ") +
-      "]"
-    );
-  });
-
-  if (testList.length === 0) {
-    var errorMessage =
-      testFilePaths.length > 0
-        ? "I couldn't find any exposed values of type Test in files matching \"" +
-          testFilePaths.toString() +
-          '"\n\nMaybe try running elm-test with no arguments?'
-        : "I couldn't find any exposed values of type Test in any *.elm files in the tests/ directory of your project's root directory.\n\nTo generate some initial tests to get things going, run elm-test init";
-
-    console.error(errorMessage);
-    process.exit(1);
-  }
-
-  function sanitizedToString(str) {
-    return '"' + str.replace(/\\/g, "\\\\").replace(/"/g, '\\"') + '"';
-  }
-
-  function getReportCode() {
-    switch (args.report) {
-      case "json":
-        return "JsonReport";
-      case "junit":
-        return "JUnitReport";
-      default:
-        if (supportsColor) {
-          return "(ConsoleReport UseColor)";
-        } else {
-          return "(ConsoleReport Monochrome)";
-        }
-    }
-  }
-
-  var fuzz = parseInt(args.fuzz);
-  var seed = parseInt(args.seed);
-
-  if (isNaN(seed)) {
-    seed = Math.floor(Math.random() * (407199254740991)) + 1000;
-  }
-
-  var opts = {
-    fuzz: isNaN(fuzz) ? "Nothing" : "Just " + fuzz,
-    seed: seed,
-    report: getReportCode(),
-    paths: testFilePaths.map(sanitizedToString).join(",")
-  };
-
-  var optsCode =
-    "{ runs = " +
-    opts.fuzz +
-    ", report = " +
-    opts.report +
-    ", seed = " +
-    opts.seed +
-    ", processes = " +
-    processes +
-    ", paths = [" +
-    opts.paths +
-    "]}";
-
-  var testFileBody = [
-    imports.join("\n"),
-    "",
-    "import Test.Reporter.Reporter exposing (Report(..))",
-    "import Console.Text exposing (UseColor(..))",
-    "import Test.Runner.Node",
-    "import Test",
-    "",
-    "main : Test.Runner.Node.TestProgram",
-    "main =",
-    "    [ " + testList.join(", ") + " ]",
-    "        |> Test.concat",
-    "        |> Test.Runner.Node.run " + optsCode
-  ].join("\n");
-
-  // Generate a filename that incorporates the hash of file contents.
-  // This way, if you run e.g. `elm-test Foo.elm` and then `elm-test Bar.elm`
-  // and then re-run `elm-test Foo.elm` we still have a cached `Main` for
-  // `Foo.elm` (assuming none of its necessary imports have changed - and
-  // why would they?) so we don't have to recompile it.
-  var salt = murmur.murmur3(testFileBody);
-  var moduleName = "Main" + salt;
-  var mainPath = path.join(generatedSrc, "Test", "Generated");
-  var mainFile = path.join(mainPath, moduleName + ".elm");
-  // We'll be putting the generated Main in something like this:
-  //
-  // my-project-name/elm-stuff/generated-code/elm-community/elm-test/src/Test/Generated/Main123456.elm
-  var testFileContents = [
-    "module Test.Generated." + moduleName + " exposing (main)",
-    testFileBody
-  ].join("\n\n");
-
-  // Make sure src/Test/Generated/ exists so we can write the file there.
-  fs.mkdirpSync(mainPath);
-
-  // Always write the file, in order to update its timestamp. This is important,
-  // because if we run `elm-make Main123456.elm` and that file's timestamp did
-  // not change, elm-make will short-circuit and not recompile *anything* - even
-  // if some of Main's dependencies (such as an individual test file) changed.
-  fs.writeFileSync(mainFile, testFileContents);
-
-  return runTests(generatedCodeDir, mainFile);
-}
-
-function getWatchPaths(projectRootDir) {
+function getWatchPaths(projectRootDir/*:string*/) {
   var watchedSourcePaths;
   var elmJson = fs.readJsonSync(path.join(projectRootDir, "elm.json"), "utf8");
   if (elmJson["type"] === "package") {
@@ -390,6 +262,14 @@ function infoLog(msg) {
   }
 }
 
+function getProjectRootDir(filePaths) {
+  return path.resolve(Runner.findNearestElmPackageDir(filePaths));
+}
+
+function getTestRootDir(projectRootDir/*:string*/) {
+  return path.resolve(path.join(projectRootDir, "tests"));
+}
+
 if (args._[0] === "make") {
   const files = args._.slice(1);
 
@@ -399,7 +279,9 @@ if (args._[0] === "make") {
   // Ergo, globify all the arguments we receive.
   const fileGlobs = files.length > 0 ? files : [];
   const testFilePaths = resolveGlobs(files);
-  const generatedCodeDir = Generate.generateElmJson(pathToElmBinary, files)[0];
+  const projectRootDir = getProjectRootDir(testFilePaths);
+  const testRootDir = getTestRootDir(projectRootDir);
+  const generatedCodeDir = Generate.generateElmJson(projectRootDir, testRootDir, pathToElmBinary, testFilePaths)[0];
 
   Compile.compileAll(testFilePaths, generatedCodeDir, args.verbose, pathToElmBinary, args.report).then(function() {
     process.exit(0);
@@ -426,7 +308,9 @@ if (args._[0] === "make") {
     process.exit(1);
   }
 
-  const returnValues = Generate.generateElmJson(pathToElmBinary, fileGlobs);
+  const projectRootDir = getProjectRootDir(testFilePaths);
+  const testRootDir = getTestRootDir(projectRootDir);
+  const returnValues = Generate.generateElmJson(projectRootDir, testRootDir, pathToElmBinary, testFilePaths);
   const generatedCodeDir = returnValues[0];
   const generatedSrc = returnValues[1];
   const sourceDirs = returnValues[2];
@@ -445,12 +329,17 @@ if (args._[0] === "make") {
         .then(function(runnableTests) {
           process.chdir(generatedCodeDir);
 
-          return generateAndRunTests(
-            runnableTests,
+          const mainFile = Generate.generateMainModule(
+            parseInt(args.fuzz),
+            parseInt(args.seed),
+            args.report,
             testFilePaths,
-            generatedCodeDir,
-            generatedSrc
+            runnableTests,
+            generatedSrc,
+            processes
           );
+
+          return runTests(generatedCodeDir, mainFile);
         })
         .catch(function(err) {
             console.error(err);
