+++ lib/runner.js
@@ -100,7 +100,7 @@ function findTests(
   testFilePaths /*: Array<string>*/,
   sourceDirs /*: Array<string>*/,
   verbose /*: boolean*/
-)/*:Promise<void>*/ {
+)/*:Promise<Array<Object>>*/ {
   return new Promise(function(resolve, reject) {
     function finish() {
       var process = spawn(readElmiPath, [], { cwd: elmPackageJsonPath });
